knownCH codes
made a series of changes.
there's two versions of knownCH, one with an FSM, and one without.

I realized the FSM is needed for this module because the node needs to receive information first, and then decide based on the information who becomes cluster head and who doesn't.

So a second version is made with the FSM in mind.

If I try to incorporate the FSM with the existing always block, magugulo sya, so I made a second one instead.