added stuff
added logic to not record cluster head information when MY_NODE_ID is equal to incoming CH_ID information.

The code currently does not consider the scenario when you are the cluster head. Since this is for storing and deciding on related cluster head information, you can create logic for selecting your nexthop cluster head in the reward block, which is outside of knownCH.

added more details sa testbench