reward updates
an addt'l signal is added pertaining to signal strength. This is currently a binary signal and this signla puts an influence in a hardware block that's outside the scope of this capstone project. The goal of the next testbench is to show waveforms for CHE and INV on the perspective of a non-CH node.